Refactor backend handlers
